refactor: 리팩토링 백로그 후속 정리
